+++ syne_tune/optimizer/schedulers/searchers/bayesopt/datatypes/common.py
@@ -10,7 +10,7 @@
 # on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
 # express or implied. See the License for the specific language governing
 # permissions and limitations under the License.
-from typing import Union, Dict, Optional, Callable, List
+from typing import Union, Dict, Optional, List
 from dataclasses import dataclass
 import numpy as np
 
@@ -24,19 +24,6 @@ INTERNAL_COST_NAME = "cost_metric"
 
 def dictionarize_objective(x):
     return {INTERNAL_METRIC_NAME: x}
-
-
-Hyperparameter = Union[str, int, float]
-
-Configuration = Dict[str, Hyperparameter]
-
-
-# Type of `filter_observed_data`, which is (optionally) used to filter the
-# observed data in `TuningJobState.trials_evaluations` when determining
-# the best config (incumbent) or the exclusion list. One use case is
-# warm-starting, where the observed data can come from a number of tasks, only
-# one of which is active.
-ConfigurationFilter = Callable[[Configuration], bool]
 
 
 MetricValues = Union[float, Dict[str, float]]
