+++ syne_tune/optimizer/schedulers/searchers/bayesopt/datatypes/config_ext.py
@@ -17,9 +17,7 @@ from syne_tune.config_space import randint
 from syne_tune.optimizer.schedulers.searchers.utils.hp_ranges import (
     HyperparameterRanges,
 )
-from syne_tune.optimizer.schedulers.searchers.bayesopt.datatypes.common import (
-    Configuration,
-)
+from syne_tune.optimizer.schedulers.searchers.utils.common import Configuration
 
 RESOURCE_ATTR_PREFIX = "RESOURCE_ATTR_"
