+++ syne_tune/optimizer/schedulers/searchers/bayesopt/models/cost/cost_model.py
@@ -13,9 +13,7 @@
 from typing import List
 from dataclasses import dataclass
 
-from syne_tune.optimizer.schedulers.searchers.bayesopt.datatypes.common import (
-    Configuration,
-)
+from syne_tune.optimizer.schedulers.searchers.utils.common import Configuration
 from syne_tune.optimizer.schedulers.searchers.bayesopt.datatypes.tuning_job_state import (
     TuningJobState,
 )
